New Trying

+++ src/task1.cpp
@@ -7,7 +7,7 @@ int comp1(void* a, void* b) {
 }
 
 int comp2(void* a, void* b) {
-	return ((*(double*)a) - (*(double*)b));
+	return static_cast<int>((*(double*)a) - (*(double*)b));
 }
 
 int comp3(void* a, void* b) {
@@ -15,7 +15,7 @@ int comp3(void* a, void* b) {
 }
 
 int comp4(void* a, void* b) {
-	return (strlen((const char*)a) - strlen((const char*)b));
+	return static_cast<int>(strlen((const char*)a) - strlen((const char*)b));
 }
 
 int comp5(void* a, void* b) {
@@ -35,6 +35,6 @@ int comp5(void* a, void* b) {
 }
 
 int comp6(void* a, void* b) {
-	return ( (((Person*)a)->age) - (((Person*)b)->age) );
+	return static_cast<int>( ((Person*)a)->age - ((Person*)b)->age );
 }
